==> input_datasets/2_cmems_join_surface_profiles.ipynb <==
import os 
import xarray as xr
import pandas as pd 
from datetime import datetime, timedelta

# Loads one day of data
def add_surface_data(data_path, data_surface_path, date, time_window):    

    sst_file_path = os.path.join(data_surface_path, f'{date}_sst.nc')
    sst = xr.open_dataset(sst_file_path)

    sss_file_path = os.path.join(data_surface_path, f'{date}_sss.nc')
    sss = xr.open_dataset(sss_file_path)

    uo_file_path = os.path.join(data_surface_path, f'{date}_uo.nc')
    uo = xr.open_dataset(uo_file_path)

    vo_file_path = os.path.join(data_surface_path, f'{date}_vo.nc')
    vo = xr.open_dataset(vo_file_path)

    mlt_file_path = os.path.join(data_surface_path, f'{date}_mlt.nc')
    mlt = xr.open_dataset(mlt_file_path)

    ssh_file_path = os.path.join(data_surface_path, f'{date}_ssh.nc')
    ssh = xr.open_dataset(ssh_file_path)
    
    parsed_date = datetime.strptime(date, "%Y%m%d")

    start_date = parsed_date - timedelta(days=time_window)
    end_date = parsed_date + timedelta(days=time_window)

    # load all files in +-window
    dates = pd.date_range(start=start_date, end=end_date).strftime("%Y%m%d")

    dfs = []
    
    for date in dates:
        file_path = f"{data_path}/{date}.feather"
        if os.path.exists(file_path):
            df = pd.read_feather(file_path)
            dfs.append(df)
        else:
            print(f"File not found for date: {date}")

    if dfs:
        profiles = pd.concat(dfs, ignore_index=True)
    else:
        return None
        
    # get unique lat-lon pairs to boost efficiency
    unique_coordinates_df = profiles[['LATITUDE', 'LONGITUDE']].drop_duplicates(subset=['LATITUDE', 'LONGITUDE'])
    
    # i get the sss and assign it. Same with sst
    unique_coordinates_df['SSS'] = unique_coordinates_df.apply(
        lambda row: sss.so_oras.sel(latitude=row['LATITUDE'], longitude=row['LONGITUDE'], method='nearest').values,
        axis=1
    )
    
    unique_coordinates_df['SST'] = unique_coordinates_df.apply(
        lambda row: sst.thetao_oras.sel(latitude=row['LATITUDE'], longitude=row['LONGITUDE'], method='nearest').values,
        axis=1
    )

    unique_coordinates_df['VO'] = unique_coordinates_df.apply(
        lambda row: vo.vo_oras.sel(latitude=row['LATITUDE'], longitude=row['LONGITUDE'], method='nearest').values,
        axis=1
    )  

    unique_coordinates_df['UO'] = unique_coordinates_df.apply(
        lambda row: uo.uo_oras.sel(latitude=row['LATITUDE'], longitude=row['LONGITUDE'], method='nearest').values,
        axis=1
    )  
    
    unique_coordinates_df['SSH'] = unique_coordinates_df.apply(
        lambda row: ssh.zos_oras.sel(latitude=row['LATITUDE'], longitude=row['LONGITUDE'], method='nearest').values,
        axis=1
    )  

    unique_coordinates_df['MLD'] = unique_coordinates_df.apply(
        lambda row: mlt.mlotst_oras.sel(latitude=row['LATITUDE'], longitude=row['LONGITUDE'], method='nearest').values,
        axis=1
    )  
    
    unique_coordinates_df['SSS'] = unique_coordinates_df['SSS'].astype(float)
    unique_coordinates_df['SST'] = unique_coordinates_df['SST'].astype(float)    
    unique_coordinates_df['VO'] = unique_coordinates_df['VO'].astype(float)
    unique_coordinates_df['UO'] = unique_coordinates_df['UO'].astype(float)    
    unique_coordinates_df['SSH'] = unique_coordinates_df['SSH'].astype(float)
    unique_coordinates_df['MLD'] = unique_coordinates_df['MLD'].astype(float)
    
    profiles = pd.merge(profiles, unique_coordinates_df, on=['LATITUDE', 'LONGITUDE'], how='left')
    profiles.drop(columns=['index', 'PRES'], inplace=True)
    profiles['DATE'] = pd.to_datetime(date)
    profiles['DATE'] = profiles['DATE'].astype(int) // 10**9
    
    return profiles

input_data = '/storage/model_collocated_simple'
input_data_surface = '/storage/model_surface'
output_data = '/storage/model_collocated_10d'
start_date = '20221015'
end_date = '20221231'

# Get all feather files (named by date)
files = [f for f in os.listdir(input_data) if f.endswith('.feather') and os.path.isfile(os.path.join(input_data, f))]

file_names = [os.path.splitext(f)[0] for f in files]
for file in file_names:
    if ((file >= start_date) and (file <= end_date)):
        print(file)
        ds = add_surface_data(input_data, input_data_surface, file, 5) 
        ds.to_feather(f'{output_data}/{file}.feather')

import pandas as pd

pd.read_feather('/storage/model_collocated_10d/2010/20100101.feather')
